+++ openapi_builder/converters/defaults/__init__.py
@@ -1,3 +1,1 @@
 """Default converters are designed for processing the default parameter."""
-
-PARAMETERS = []