+++ openapi_builder/converters/parameter/__init__.py
@@ -1,3 +1,1 @@
 """Parameter converters are designed for processing the URL parameter."""
-
-PARAMETERS = []